refactored reactor_element out
