+++ include/reactor-cpp/action.hh
@@ -16,6 +16,7 @@
 #include "reactor.hh"
 #include "time_barrier.hh"
 #include "value_ptr.hh"
+#include "reactor_element.hh"
 
 #include <condition_variable>
 #include <map>
@@ -48,22 +49,13 @@ protected:
    * indicates that the tag is safe to process.
    */
   virtual auto acquire_tag(const Tag& tag, std::unique_lock<std::mutex>& lock,
-                           const std::function<bool(void)>& abort_waiting) -> bool {
-    reactor_assert(!logical_);
-    reactor_assert(lock.owns_lock());
-    return PhysicalTimeBarrier::acquire_tag(tag, lock, environment()->scheduler(), abort_waiting);
-  }
+                           const std::function<bool(void)>& abort_waiting) -> bool;
 
   BaseAction(const std::string& name, Reactor* container, bool logical, Duration min_delay)
       : ReactorElement(name, ReactorElement::Type::Action, container)
       , min_delay_(min_delay)
       , logical_(logical) {}
-  BaseAction(const std::string& name, Environment* environment, bool logical, Duration min_delay)
-      : ReactorElement(name, ReactorElement::Type::Action, environment)
-      , min_delay_(min_delay)
-      , logical_(logical) {
-    environment->insert_input_action(this);
-  }
+  BaseAction(const std::string& name, Environment* environment, bool logical, Duration min_delay);
 
 public:
   [[nodiscard]] auto inline triggers() const noexcept -> const auto& { return triggers_; }
@@ -150,12 +142,7 @@ public:
 
 template <class T> class PhysicalAction : public Action<T> {
 public:
-  PhysicalAction(const std::string& name, Reactor* container)
-      : Action<T>(name, container, false, Duration::zero()) {
-    // all physical actions act as input actions to the program as they can be
-    // scheduled from external threads
-    container->environment()->insert_input_action(this);
-  }
+  PhysicalAction(const std::string& name, Reactor* container);
 };
 
 template <class T> class LogicalAction : public Action<T> {
