+++ include/reactor-cpp/impl/action_impl.hh
@@ -11,6 +11,7 @@
 
 #include "../assert.hh"
 #include "../environment.hh"
+#include "../reactor.hh"
 #include <iterator>
 #include <mutex>
 
@@ -99,6 +100,13 @@ template <class T> void Action<T>::setup() noexcept {
 template <class T> void Action<T>::cleanup() noexcept {
   BaseAction::cleanup();
   value_ptr_ = nullptr;
+}
+
+template <class T>
+PhysicalAction<T>::PhysicalAction(const std::string& name, reactor::Reactor* container): Action<T>(name, container, false, Duration::zero()) {
+  // all physical actions act as input actions to the program as they can be
+  // scheduled from external threads
+  this->environment()->insert_input_action(this);
 }
 
 } // namespace reactor
